chore: create families package
