+++ packages/client/src/index.ts
@@ -1,6 +1,7 @@
 /* istanbul ignore file */
 import SDK from "./LedgerLivePlatformSDK";
 
+export * from "@ledgerhq/wallet-api-families";
 export * from "@ledgerhq/wallet-api-transports";
 export * from "@ledgerhq/wallet-api-types";
 export * from "./serializers";
