+++ packages/families/.eslintrc.js
@@ -1,0 +1,14 @@
+module.exports = {
+  env: {
+    es2021: true,
+    node: true,
+  },
+  parserOptions: {
+    project: ["./tsconfig.json"],
+    tsconfigRootDir: __dirname,
+  },
+  rules: {
+    "no-underscore-dangle": "off",
+    "class-methods-use-this": "warn",
+  },
+};
