+++ packages/families/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ledgerhq/wallet-api-transports",
+  "name": "@ledgerhq/wallet-api-families",
   "version": "0.1.0",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -19,8 +19,6 @@
     "typescript": "^4.5.2"
   },
   "dependencies": {
-    "@ledgerhq/wallet-api-logger": "workspace:*",
-    "@ledgerhq/wallet-api-types": "workspace:*",
     "bignumber.js": "^9.0.2"
   }
 }
