+++ packages/families/src/tron/serializer.ts
@@ -1,0 +1,38 @@
+import BigNumber from "bignumber.js";
+import type { RawTronTransaction, TronTransaction } from "./types";
+
+export const serializeTronTransaction = ({
+  amount,
+  recipient,
+  family,
+  mode,
+  resource,
+  duration,
+}: TronTransaction): RawTronTransaction => {
+  return {
+    amount: amount.toString(),
+    recipient,
+    family,
+    mode,
+    resource,
+    duration,
+  };
+};
+
+export const deserializeTronTransaction = ({
+  amount,
+  recipient,
+  family,
+  mode,
+  resource,
+  duration,
+}: RawTronTransaction): TronTransaction => {
+  return {
+    amount: new BigNumber(amount),
+    recipient,
+    family,
+    mode,
+    resource,
+    duration,
+  };
+};
