+++ packages/logger/package.json
@@ -13,7 +13,7 @@
   },
   "devDependencies": {
     "@tsconfig/node16-strictest-esm": "^1.0.3",
-    "@types/node": "^18.7.9",
+    "@types/node": "^18.7.18",
     "eslint": "^7.32.0",
     "tsup": "^6.2.3",
     "typescript": "^4.5.2"
