--- packages/types/src/families/cosmos/types.ts
@@ -1,30 +1,0 @@
-import type BigNumber from "bignumber.js";
-import type {
-  FAMILIES,
-  RawTransactionCommon,
-  TransactionCommon,
-} from "../../index";
-
-export type CosmosOperationMode =
-  | "send"
-  | "delegate"
-  | "undelegate"
-  | "redelegate"
-  | "claimReward"
-  | "claimRewardCompound";
-
-export interface CosmosTransaction extends TransactionCommon {
-  readonly family: FAMILIES.COSMOS;
-  mode: CosmosOperationMode;
-  fees?: BigNumber;
-  gas?: BigNumber;
-  memo?: string;
-}
-
-export interface RawCosmosTransaction extends RawTransactionCommon {
-  readonly family: FAMILIES.COSMOS;
-  mode: CosmosOperationMode;
-  fees?: string;
-  gas?: string;
-  memo?: string;
-}
