--- packages/types/src/families/polkadot/types.ts
@@ -1,31 +1,0 @@
-import type BigNumber from "bignumber.js";
-import type {
-  FAMILIES,
-  RawTransactionCommon,
-  TransactionCommon,
-} from "../../index";
-
-export type PolkadotOperationMode =
-  | "send"
-  | "bond"
-  | "unbond"
-  | "rebond"
-  | "withdrawUnbonded"
-  | "setController"
-  | "nominate"
-  | "chill"
-  | "claimReward";
-
-export interface PolkadotTransaction extends TransactionCommon {
-  readonly family: FAMILIES.POLKADOT;
-  mode: PolkadotOperationMode;
-  fee?: BigNumber;
-  era?: number;
-}
-
-export interface RawPolkadotTransaction extends RawTransactionCommon {
-  readonly family: FAMILIES.POLKADOT;
-  mode: PolkadotOperationMode;
-  fee?: string;
-  era?: number;
-}
