--- packages/types/src/families/ripple/serializer.ts
@@ -1,34 +1,0 @@
-import BigNumber from "bignumber.js";
-import type { RawRippleTransaction, RippleTransaction } from "../../index";
-
-export const serializeRippleTransaction = ({
-  family,
-  fee,
-  tag,
-  amount,
-  recipient,
-}: RippleTransaction): RawRippleTransaction => {
-  return {
-    family,
-    amount: amount.toString(),
-    recipient,
-    fee: fee ? fee.toString() : undefined,
-    tag,
-  };
-};
-
-export const deserializeRippleTransaction = ({
-  family,
-  fee,
-  tag,
-  amount,
-  recipient,
-}: RawRippleTransaction): RippleTransaction => {
-  return {
-    family,
-    amount: new BigNumber(amount),
-    recipient,
-    fee: fee ? new BigNumber(fee) : undefined,
-    tag,
-  };
-};
