--- packages/types/src/families/stellar/serializer.ts
@@ -1,38 +1,0 @@
-import BigNumber from "bignumber.js";
-import type { RawStellarTransaction, StellarTransaction } from "../../index";
-
-export const serializeStellarTransaction = ({
-  amount,
-  recipient,
-  family,
-  fees,
-  memoType,
-  memoValue,
-}: StellarTransaction): RawStellarTransaction => {
-  return {
-    amount: amount.toString(),
-    recipient,
-    family,
-    fees: fees ? fees.toString() : undefined,
-    memoType,
-    memoValue,
-  };
-};
-
-export const deserializeStellarTransaction = ({
-  amount,
-  recipient,
-  family,
-  fees,
-  memoType,
-  memoValue,
-}: RawStellarTransaction): StellarTransaction => {
-  return {
-    amount: new BigNumber(amount),
-    recipient,
-    family,
-    fees: fees ? new BigNumber(fees) : undefined,
-    memoType,
-    memoValue,
-  };
-};
