--- packages/types/src/families/tron/serializer.ts
@@ -1,38 +1,0 @@
-import BigNumber from "bignumber.js";
-import type { RawTronTransaction, TronTransaction } from "../../index";
-
-export const serializeTronTransaction = ({
-  amount,
-  recipient,
-  family,
-  mode,
-  resource,
-  duration,
-}: TronTransaction): RawTronTransaction => {
-  return {
-    amount: amount.toString(),
-    recipient,
-    family,
-    mode,
-    resource,
-    duration,
-  };
-};
-
-export const deserializeTronTransaction = ({
-  amount,
-  recipient,
-  family,
-  mode,
-  resource,
-  duration,
-}: RawTronTransaction): TronTransaction => {
-  return {
-    amount: new BigNumber(amount),
-    recipient,
-    family,
-    mode,
-    resource,
-    duration,
-  };
-};
