--- packages/types/src/families/tron/types.ts
@@ -1,28 +1,0 @@
-import type {
-  FAMILIES,
-  RawTransactionCommon,
-  TransactionCommon,
-} from "../../index";
-
-export type TronOperationMode =
-  | "send"
-  | "freeze"
-  | "unfreeze"
-  | "vote"
-  | "claimReward";
-
-export type TronResource = "BANDWIDTH" | "ENERGY";
-
-export interface TronTransaction extends TransactionCommon {
-  readonly family: FAMILIES.TRON;
-  mode: TronOperationMode;
-  resource?: TronResource;
-  duration?: number;
-}
-
-export interface RawTronTransaction extends RawTransactionCommon {
-  readonly family: FAMILIES.TRON;
-  mode: TronOperationMode;
-  resource?: TronResource;
-  duration?: number;
-}
